fix: 모달 새로고침 페이지 수정
