+++ src/app/(routers)/(todo)/goals/todos/new/page.tsx
@@ -1,3 +1,9 @@
-export default function newTodo() {
-  return <div>newTodo</div>;
+import NewForm from '../../../@modal/(.)goals/todos/new/_components/newForm';
+
+export default function NewTodoPage() {
+  return (
+    <div className="mx-auto max-w-lg py-10">
+      <NewForm />
+    </div>
+  );
 }
